# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Global App Admin Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /apps/{appId}/ledger:
    get:
      operationId: getLedger
      tags: [ledger]
      summary: Financial ledger (balances, reconciliation, recent transactions) for an app
      parameters:
        - $ref: "#/components/parameters/AppId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LedgerReport" }
  /apps/{appId}/ledger/entries:
    get:
      operationId: listLedgerEntries
      tags: [ledger]
      summary: List journal entries for an app, newest first
      parameters:
        - $ref: "#/components/parameters/AppId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LedgerJournalEntry" }
    post:
      operationId: postLedgerEntry
      tags: [ledger]
      summary: Post a new balanced double-entry journal entry
      parameters:
        - $ref: "#/components/parameters/AppId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PostLedgerEntryRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LedgerJournalEntry" }
        "400":
          description: Invalid entry (unbalanced, unknown account, or bad amount)
        "404":
          description: App not found
  /apps/{appId}/ledger/sales:
    post:
      operationId: ingestLedgerSale
      tags: [ledger]
      summary: Ingest a platform sale, netting out the platform fee into the ledger
      description: >-
        Records a sale as a balanced pair of journal entries: gross revenue
        recognized and the platform fee (Apple 30%, Google 15%, Stripe 3%) booked
        as an expense, leaving net cash correct. Idempotent on (source, externalRef).
      parameters:
        - $ref: "#/components/parameters/AppId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IngestSaleRequest" }
      responses:
        "201":
          description: Sale recorded (or already recorded — idempotent)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IngestSaleResult" }
        "400":
          description: Invalid sale (bad amount, unknown source, or missing accounts)
        "404":
          description: App not found
  /apps/{appId}/ledger/stripe/sync:
    post:
      operationId: syncStripeSales
      tags: [ledger]
      summary: Import live Stripe charges into the app's ledger with their actual fees
      description: >-
        Pulls succeeded Stripe charges and records each as a balanced sale,
        booking Stripe's actual reported per-transaction fee as the platform-fee
        expense. Idempotent on the Stripe charge id, so re-running is safe.
        Enabled for GrailBabe only.
      parameters:
        - $ref: "#/components/parameters/AppId"
      responses:
        "200":
          description: Sync completed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StripeSyncResult" }
        "400":
          description: Stripe sync is not enabled for this app
        "404":
          description: App not found
        "503":
          description: Stripe is not connected
  /apps/{appId}/ledger/reconcile:
    post:
      operationId: reconcileLedger
      tags: [ledger]
      summary: Run reconciliation for an app's ledger and persist the result
      parameters:
        - $ref: "#/components/parameters/AppId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LedgerReconciliation" }
        "404":
          description: App not found
...
components:
  parameters:
    AppId:
      name: appId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    LedgerAccountType:
      type: string
      enum: [asset, liability, equity, revenue, expense]
    LedgerAccount:
      type: object
      required: [code, name, type, balance]
      properties:
        code: { type: string, description: Chart-of-accounts code }
        name: { type: string }
        type: { $ref: "#/components/schemas/LedgerAccountType" }
        balance: { type: number, description: Current balance in the ledger currency }
    LedgerReconciliation:
      type: object
      required: [status, lastReconciledAt, unreconciledCount, unreconciledAmount]
      properties:
        status: { type: string, enum: [reconciled, pending, discrepancy] }
        lastReconciledAt: { type: string, format: date-time }
        unreconciledCount: { type: integer, description: Number of journal entries not yet reconciled }
        unreconciledAmount: { type: number, description: Net value of unreconciled entries }
    LedgerTransaction:
      type: object
      required: [id, postedAt, description, debitAccount, creditAccount, amount, status]
      properties:
        id: { type: string }
        postedAt: { type: string, format: date-time }
        description: { type: string }
        debitAccount: { type: string, description: Chart-of-accounts code debited }
        creditAccount: { type: string, description: Chart-of-accounts code credited }
        amount: { type: number }
        status: { type: string, enum: [posted, pending, failed] }
    LedgerEntrySource:
      type: string
      enum: [stripe, app_store, play_store, bank, manual]
    LedgerJournalEntry:
      type: object
      required: [id, postedAt, description, debitAccount, creditAccount, amount, status, source]
      properties:
        id: { type: string }
        postedAt: { type: string, format: date-time }
        description: { type: string }
        debitAccount: { type: string, description: Chart-of-accounts code debited }
        creditAccount: { type: string, description: Chart-of-accounts code credited }
        amount: { type: number }
        status: { type: string, enum: [posted, pending, failed] }
        source: { $ref: "#/components/schemas/LedgerEntrySource" }
    PostLedgerEntryRequest:
      type: object
      required: [description, debitAccount, creditAccount, amount]
      properties:
        description: { type: string, minLength: 1 }
        debitAccount: { type: string, description: Chart-of-accounts code to debit }
        creditAccount: { type: string, description: Chart-of-accounts code to credit }
        amount: { type: number, exclusiveMinimum: 0 }
        source: { $ref: "#/components/schemas/LedgerEntrySource" }
        status: { type: string, enum: [posted, pending, failed], default: posted }
        postedAt: { type: string, format: date-time }
    LedgerRevenueBySource:
      type: object
      required: [source, label, feeRate, gross, fee, net]
      properties:
        source: { $ref: "#/components/schemas/LedgerEntrySource" }
        label: { type: string, description: Human-readable source label }
        feeRate: { type: number, description: Platform fee rate (e.g. 0.3 = 30%) }
        gross: { type: number, description: Gross revenue recognized from this source }
        fee: { type: number, description: Platform fee booked as expense for this source }
        net: { type: number, description: Net cash after the platform fee (gross - fee) }
    LedgerRevenue:
      type: object
      required: [grossRevenue, platformFees, netRevenue, bySource]
      properties:
        grossRevenue: { type: number, description: Total recognized revenue before platform fees }
        platformFees: { type: number, description: Total platform fees booked as expense }
        netRevenue: { type: number, description: Net revenue after platform fees }
        bySource:
          type: array
          items: { $ref: "#/components/schemas/LedgerRevenueBySource" }
    IngestSaleRequest:
      type: object
      required: [source, grossAmount]
      properties:
        source: { $ref: "#/components/schemas/LedgerEntrySource" }
        grossAmount: { type: number, exclusiveMinimum: 0, description: Gross sale amount before the platform fee }
        feeAmount:
          type: number
          minimum: 0
          description: >-
            Actual platform fee for this sale. When provided it overrides the
            flat schedule rate — used by live feeds (e.g. Stripe) that report the
            real per-transaction fee. Must not exceed grossAmount.
        description: { type: string, minLength: 1 }
        externalRef: { type: string, description: External transaction id for idempotent ingestion }
        status: { type: string, enum: [posted, pending, failed], default: posted }
        postedAt: { type: string, format: date-time }
    IngestSaleResult:
      type: object
      required: [source, feeRate, gross, fee, net, entries]
      properties:
        source: { $ref: "#/components/schemas/LedgerEntrySource" }
        created: { type: boolean, description: True when new legs were recorded; false on an idempotent replay }
        feeRate: { type: number, description: Effective platform fee rate booked (e.g. 0.3 = 30%) }
        gross: { type: number }
        fee: { type: number }
        net: { type: number, description: Net cash after the platform fee }
        entries:
          type: array
          description: The journal entries created (or already present) for this sale.
          items: { $ref: "#/components/schemas/LedgerJournalEntry" }
    StripeSyncSkip:
      type: object
      required: [chargeId, reason]
      properties:
        chargeId: { type: string }
        reason: { type: string }
    StripeSyncResult:
      type: object
      required: [fetched, imported, alreadyRecorded, skipped, gross, fee, net, skips]
      properties:
        fetched: { type: integer, description: Total Stripe charges examined }
        imported: { type: integer, description: Charges newly recorded in the ledger this run }
        alreadyRecorded: { type: integer, description: Charges already present (idempotent replays) }
        skipped: { type: integer, description: Charges skipped (not succeeded, non-USD, or errored) }
        gross: { type: number, description: Total gross of all valid Stripe sales in the ledger }
        fee: { type: number, description: Total actual Stripe fees booked }
        net: { type: number, description: Total net cash after Stripe fees }
        skips:
          type: array
          description: Per-charge skip reasons, capped for readability.
          items: { $ref: "#/components/schemas/StripeSyncSkip" }
    LedgerReport:
      type: object
      required: [currency, totalBalance, accounts, reconciliation, transactions, revenue]
      properties:
        currency: { type: string }
        totalBalance: { type: number, description: Net of asset balances (settlement position) }
        accounts:
          type: array
          items: { $ref: "#/components/schemas/LedgerAccount" }
        reconciliation: { $ref: "#/components/schemas/LedgerReconciliation" }
        transactions:
          type: array
          description: Most recent journal entries, newest first.
          items: { $ref: "#/components/schemas/LedgerTransaction" }
        revenue: { $ref: "#/components/schemas/LedgerRevenue" }
...